Update RigolDS1054Z.py
This class provides the functionality and contains the framework for data acquisition and remote scope control. Not all functions are called in given Run scripts; extra functions are provided to give user room for other uses (such as simple control of the scope), or further development.  However, if using this code as a library simply for scope control, note there are many self variables established during initialization, not represented in the _init_ function, which are referred to by many functions.